feat: Allow integer ratings for list items
Updates the `list_items` table schema to accept integer ratings. The `App.tsx` file is updated to round any incoming floating-point ratings to the nearest integer before saving them, ensuring consistency.

+++ lib/supabaseClient.ts
@@ -409,7 +409,7 @@ create table public.list_items (
   user_id uuid not null,
   show_id bigint not null,
   status character varying(50) not null,
-  rating integer null,
+  rating numeric null,
   progress integer null default 0,
   notes text null,
   start_date timestamp with time zone null,
